Move octokit code inside src/dev/github
