+++ scripts/get_serverless_release_sha.js
@@ -1,0 +1,23 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+require('../src/setup_node_env');
+var getServerlessReleaseSha =
+  require('../src/dev/github/get_serverless_release_sha').getServerlessReleaseSha;
+
+getServerlessReleaseSha().then(
+  function (sha) {
+    console.log(sha);
+    process.exit(0);
+  },
+  function (reason) {
+    console.error(reason);
+    process.exit(1);
+  }
+);
